+++ packages/release/src/releases/index.ts
@@ -3,6 +3,7 @@ import { default as Release_0_15_0 } from "./release_0_15_0.mdx";
 import { default as Release_0_16_0 } from "./release_0_16_0.mdx";
 import { default as Release_0_17_0 } from "./release_0_17_0.mdx";
 import { default as Release_0_18_0 } from "./release_0_18_0.mdx";
+import { default as Release_0_19_0 } from "./release_0_19_0.mdx";
 
 export const releaseMDXComponents = {
   Release_0_14_0: Release_0_14_0,
@@ -10,4 +11,5 @@ export const releaseMDXComponents = {
   Release_0_16_0: Release_0_16_0,
   Release_0_17_0: Release_0_17_0,
   Release_0_18_0: Release_0_18_0,
+  Release_0_19_0: Release_0_19_0,
 };